add search bar to recovery pages
