+++ BrightID/src/routes/Groups.js
@@ -1,9 +1,8 @@
 // @flow
 
-import React, { useRef, useEffect } from 'react';
-import { Animated, Text } from 'react-native';
+import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
-import { useSelector } from 'react-redux';
+import { useTranslation } from 'react-i18next';
 import { getGroupName } from '@/utils/groups';
 import { DEVICE_TYPE } from '@/utils/deviceConstants';
 import GroupsScreen from '@/components/GroupsScreens/GroupsScreen';
@@ -13,47 +12,21 @@ import NewGroupScreen from '@/components/GroupsScreens/NewGroups/NewGroupScreen'
 import GroupInfoScreen from '@/components/GroupsScreens/NewGroups/GroupInfoScreen';
 import MembersScreen from '@/components/GroupsScreens/Members/MembersScreen';
 import InviteListScreen from '@/components/GroupsScreens/Members/InviteListScreen';
-import { headerOptions, headerTitleStyle, NavHome } from './helpers';
-import { useTranslation } from 'react-i18next';
+import { headerOptions, NavHome, AnimatedHeaderTitle } from './helpers';
 
 const Stack = createStackNavigator();
-
-const HeaderTitle = ({ title }) => {
-  const searchOpen = useSelector(
-    (state) => state.groups.searchOpen || state.connections.searchOpen,
-  );
-  const fadeAnim = useRef(new Animated.Value(1)).current;
-  useEffect(() => {
-    Animated.timing(fadeAnim, {
-      toValue: searchOpen ? 0 : 1,
-      useNativeDriver: true,
-      duration: 600,
-    }).start();
-  }, [fadeAnim, searchOpen]);
-
-  return (
-    <Animated.View style={{ opacity: fadeAnim }}>
-      <Text style={headerTitleStyle}>{title}</Text>
-    </Animated.View>
-  );
-};
-
-const HeaderTitleI18N = ({ i18key }) => {
-  const { t } = useTranslation();
-  return (<HeaderTitle title={t(i18key)}/>);
-};
 
 const groupsOptions = {
   ...headerOptions,
   headerRight: () => <SearchGroups />,
   headerLeft: () => <NavHome />,
-  headerTitle: () => <HeaderTitleI18N i18key="groups.header.groups" />,
+  headerTitle: () => <AnimatedHeaderTitle i18key="groups.header.groups" />,
 };
 
 const newGroupOptions = {
   ...headerOptions,
   headerRight: () => <SearchConnections />,
-  headerTitle: () => <HeaderTitleI18N i18key="groups.header.newGroup" />,
+  headerTitle: () => <AnimatedHeaderTitle i18key="groups.header.newGroup" />,
 };
 
 const membersScreenOptions = ({ navigation, route }) => {
@@ -91,7 +64,7 @@ const Groups = () => {
       <Stack.Screen
         name="Members"
         component={MembersScreen}
-        options={membersScreenOptions}  
+        options={membersScreenOptions}
       />
       <Stack.Screen
         name="InviteList"
@@ -99,7 +72,7 @@ const Groups = () => {
         options={{ ...headerOptions, title: t('groups.header.inviteList') }}
       />
     </>
-  )
+  );
 };
 
 export default Groups;
