+++ BrightID/src/routes/Home.js
@@ -7,7 +7,7 @@ import {
   View,
 } from 'react-native';
 import { useSelector } from 'react-redux';
-import { useTranslation } from 'react-i18next';
+// import { useTranslation } from 'react-i18next';
 import { INVITE_ACTIVE } from '@/utils/constants';
 import { DEVICE_LARGE } from '@/utils/deviceConstants';
 import { createSelector } from '@reduxjs/toolkit';
@@ -18,14 +18,12 @@ import {
   pendingConnection_states,
   selectAllPendingConnections,
 } from '@/components/PendingConnectionsScreens/pendingConnectionSlice';
-import RecoveringConnectionScreen from '@/components/Recovery/RecoveringConnectionScreen';
 import {
   toggleDrawer,
   resetHome,
   resetNotifications,
 } from '@/NavigationService';
 import menuBar from '@/static/menu_bar.svg';
-import { headerOptions } from './helpers';
 import { HomeDrawer } from './HomeDrawer';
 
 /** SELECTORS */
@@ -139,28 +137,19 @@ const homeScreenOptions = {
   headerTransparent: true,
 };
 
-
 /** SCREENS */
 
 const Stack = createStackNavigator();
 
 const Home = () => {
-  /* /!\ Moved 'RecoveringConnection' options to the component in order to access to useTranslation() */ 
-  const { t } = useTranslation();
+  /* /!\ Moved 'RecoveringConnection' options to the component in order to access to useTranslation() */
+  // const { t } = useTranslation();
   return (
     <>
       <Stack.Screen
         name="Home"
         component={HomeDrawer}
         options={homeScreenOptions}
-      />
-      <Stack.Screen
-        name="RecoveringConnection"
-        component={RecoveringConnectionScreen}
-        options={{
-          ...headerOptions,
-          title: t('restore.header.accountRecovery')
-        }}
       />
     </>
   );
