+++ BrightID/src/routes/Notifications.js
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
-import { useTranslation } from 'react-i18next';
+// import { useTranslation } from 'react-i18next';
 import NotificationsScreen from '@/components/Notifications/NotificationsScreen';
-import BackupScreen from '@/components/Recovery/BackupScreen';
 import { headerOptions, NavHome } from './helpers';
 
 const Stack = createStackNavigator();
@@ -13,7 +12,7 @@ const topOptions = {
 };
 
 const Notifications = () => {
-  const { t } = useTranslation();
+  // const { t } = useTranslation();
   return (
     <>
       <Stack.Screen
@@ -21,17 +20,8 @@ const Notifications = () => {
         component={NotificationsScreen}
         options={topOptions}
       />
-
-      <Stack.Screen
-        name="Backup"
-        component={BackupScreen}
-        options={{
-          ...headerOptions,
-          title: t('backup.header.backup')
-        }}
-      />
     </>
-  )
+  );
 };
 
 export default Notifications;
